+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added Tape Infrastructure to the diagram
+  - Tape Server
+  - Tape Library 
+  - Tape Vault
+- Added Service Provider to the diagram
+
 ## [0.8.7] - 2024-05-28
 
 ### Added